Исправил sequence_reader для работы с новым парсером

+++ monotonic/include/sequence_reader.h
@@ -1,8 +1,9 @@
 #ifndef SEQUENCE_READER
 #define SEQUENCE_READER
 
+#include <string.h>
 
-int readSequence(const char* filename, double** arr, int* count);
+int readSequence(const char* filename, double** arr, size_t* count);
 
 #endif
